requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.4.2
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.0
langchain==0.0.300
transformers==4.33.0
torch==2.0.1